refactor: estimatedPayment custom hook으로 분리
