+++ src/components/Cart/EstimatedPaymentBox.tsx
@@ -1,62 +1,20 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import Button from '../Common/Button';
-import { useRecoilValue } from 'recoil';
-import {
-  checkedItemAtom,
-  totalPriceSelector,
-} from '../../recoil/checkedProductData';
-import {
-  FREE_DELIVERY_THRESHOLD,
-  REWARD_POINT_RATE,
-  STANDARD_DELIVERY_FEE,
-} from '../../constants/price';
-import { hostNameAtom } from '../../recoil/hostData';
-import { orderApi } from '../../apis/orderProducts';
-import type { OrderedData } from '../../types/product';
+import useEstimatedPayment from '../../hooks/useEstimatedPayment';
 
 interface EstimatedPaymentBoxProps {
   usePoint: number;
 }
 
 const EstimatedPaymentBox = ({ usePoint }: EstimatedPaymentBoxProps) => {
-  const checkedCartProduct = useRecoilValue(checkedItemAtom);
-  const totalProductPrice = useRecoilValue(totalPriceSelector);
-  const totalDeliveryFee =
-    totalProductPrice === 0 || totalProductPrice >= FREE_DELIVERY_THRESHOLD
-      ? 0
-      : STANDARD_DELIVERY_FEE;
-  const totalPrice = totalProductPrice
-    ? totalProductPrice + totalDeliveryFee
-    : 0;
-  const rewardPoints = totalProductPrice * REWARD_POINT_RATE;
-
-  // 커스텀 훅 분리
-  const hostName = useRecoilValue(hostNameAtom);
-
-  const submitOrder = () => {
-    const cartItems = checkedCartProduct.map(
-      ({ cartItemId, quantity, product }) => {
-        return {
-          cartItemId,
-          quantity,
-          product,
-        };
-      }
-    );
-
-    const orderData: OrderedData = {
-      cartItems,
-      totalProductPrice,
-      totalDeliveryFee,
-      usePoint,
-      totalPrice,
-    };
-
-    orderApi(hostName).then((apiInstance) => {
-      return apiInstance.postOrderProduct(orderData);
-    });
-  };
+  const {
+    totalProductPrice,
+    totalDeliveryFee,
+    rewardPoints,
+    totalPrice,
+    submitOrder,
+  } = useEstimatedPayment(usePoint);
 
   return (
     <EstimatedPaymentBoxContainer>
